fixing warnings, adding missing file, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@16386 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netbuild/NBLoadedSUMOTLDef.cpp
@@ -265,7 +265,7 @@ NBLoadedSUMOTLDef::patchIfCrossingsAdded() {
         for (NBConnectionVector::const_iterator it = myControlledLinks.begin(); it != myControlledLinks.end(); it++) {
             const NBConnection& c = *it;
             if (c.getTLIndex() != NBConnection::InvalidTlIndex) {
-                assert(c.getTLIndex() < size);
+                assert(c.getTLIndex() < (int)size);
                 fromEdges[c.getTLIndex()] = c.getFrom();
                 toEdges[c.getTLIndex()] = c.getTo();
             }
